New feature: random playlists
a function was added to MusicService in order to create and set playlists
of random songs. The function makes a copy of the songCollection list and
takes songs randomly into a new song. Afterwards the list is set as the
current playlist.

The floating action button in the main activity is set to start a random
playlist on click.