New playback status: Ready
when MusicService starts it has a new status: ready. This is different from "stopped"
because service is not killed when all clients unbind. If user navigates away from
the app, the service will still be alive even if there is no playlist selected. This
prevents the app from killing the service when the screen is rotated before selecting
a playlist.